refactor(resources): move tool CRUD routes
